修复一些小问题

+++ workflow/models/store.go
@@ -1,7 +1,7 @@
 package models
 
 type Holder struct {
-	Status      int64    `json:"status"`
+	Status      int      `json:"status"`
 	ParentNodes []string `json:"parent_nodes"` //所有父节点
 	NodeId      string   `json:"node_id"`      //当前节点
 	InstanceId  string   `json:"instance_id"`  //实例ID
